single prediction manager
